Move plugin functions to own file
